requirements.txt
requests>=2.28.0
duckdb>=0.9.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
